# make `unserialize` function a little faster