--- a/src/IntegerParser.php
+++ b/src/IntegerParser.php
@@ -18,8 +18,7 @@ class IntegerParser implements ParserInterface
      */
     public function parse(Source $source)
     {
-        $source->consume('i');
-        $source->consume(':');
+        $source->consume('i:');
         $parser = new NumberLiteralParser();
         list($result, $source) = $parser->parse($source);
         $source->consume(';');