benchmark script:
```php
<?php

require 'vendor/autoload.php';

use Lavoiesl\PhpBenchmark\Benchmark;
use xKerman\Restricted;

$data = serialize([null, "key" => false, 3 => -5.6, "loooooooooooooooooooooooong", ["nested", ["array"]]]);
$benchmark = new Benchmark();

$benchmark->add('built-in', function () use ($data) { return \unserialize($data); });
$benchmark->add('restricted-unserialize', function () use ($data) { return Restricted\unserialize($data); });

$benchmark->run();
```

# before: (10 times)
```
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1891 ms    31417 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1909 ms    31717 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1903 ms    31617 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    5 ms                 0 B                
restricted-unserialize   1962 ms    39140 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1883 ms    31283 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1881 ms    31250 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1928 ms    32033 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1906 ms    31667 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    6 ms                 0 B                
restricted-unserialize   1905 ms    31650 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    8 ms                 0 B                
restricted-unserialize   1929 ms    24013 %      0 B                
```

# after: (10 times)
```
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    7 ms                 0 B                
restricted-unserialize   2420 ms    34471 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    9 ms                 0 B                
restricted-unserialize   2404 ms    26611 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    9 ms                 0 B                
restricted-unserialize   2399 ms    26556 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                   10 ms                 0 B                
restricted-unserialize   2437 ms    24270 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    9 ms                 0 B                
restricted-unserialize   2391 ms    26467 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                   10 ms                 0 B                
restricted-unserialize   2444 ms    24340 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    8 ms                 0 B                
restricted-unserialize   2486 ms    30975 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                   11 ms                 0 B                
restricted-unserialize   2346 ms    21227 %      0 B                
--------------------------------------------------------------------------------
Running tests 2000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                    5 ms                 0 B                
restricted-unserialize   1581 ms    31520 %      0 B                
--------------------------------------------------------------------------------
Running tests 3000 times.
Testing 1/2 : built-in
Testing 2/2 : restricted-unserialize

Test                        Time   Time (%)   Memory   Memory (%)   
built-in                   10 ms                 0 B                
restricted-unserialize   2626 ms    26160 %      0 B                
```